+++ CHANGELOG.md
@@ -6,11 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 
+- Parallel import (only some resources are supported. To enable use `--import.parallel=true`)
+- Don't update client if not changed
+- Don't update components config if not changed
+- Don't update realm role if not changed
 
 ### Changed
-
 
 ### Fixed